# api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Sendit Documentation",
...
        "contact": {
            "name": "Sendit Support",
            "email": "contact@sendit.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://app.sendit.ma/api/v1/",
            "description": "Sendit"
        }
    ],
    "paths": {
...
        "/movements/out": {
            "get": {
                "tags": [
                    "Sortie de stock"
                ],
                "summary": "Liste des sorties du stock",
                "description": "Retourne les sorties du stock.",
                "operationId": "3ef8d15623a93385f2c7694c88410578",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "querystring",
                        "in": "query",
                        "description": "Recherche par code, nom, téléphone, statut ou date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "RSHF04"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des sorties de stock récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des sorties de stock."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MovementListDetailOut"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Les données fournies sont invalides"
                    },
                    "403": {
                        "description": "Vous n'avez pas les autorisations nécessaires"
                    },
                    "500": {
                        "description": "Une erreur s'est produite côté serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sortie de stock"
                ],
                "summary": "Créer un mouvement de sortie de stock",
                "description": "Crée un mouvement de sortie de stock contenant produits, client, téléphone, adresse, district…",
                "operationId": "a38b79a56c2f189e273985e2f39d7297",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "products",
                                    "district_id",
                                    "name",
                                    "phone",
                                    "address"
                                ],
                                "properties": {
                                    "products": {
                                        "description": "Liste des produits au format 'CODE:QTY;CODE2:QTY'",
                                        "type": "string",
                                        "example": "PR3D06:30"
                                    },
                                    "district_id": {
                                        "description": "ID du district",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "Nom du destinataire",
                                        "type": "string",
                                        "example": "test label"
                                    },
                                    "phone": {
                                        "description": "Téléphone du destinataire",
                                        "type": "string",
                                        "example": "0698855977"
                                    },
                                    "address": {
                                        "description": "Adresse du destinataire",
                                        "type": "string",
                                        "example": "Casa"
                                    },
                                    "note": {
                                        "description": "Note éventuelle",
                                        "type": "string",
                                        "example": "Remarque optionnelle",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mouvement de sortie créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovementDetailOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Les données fournies sont invalides"
                    },
                    "403": {
                        "description": "Vous n'avez pas les autorisations nécessaires"
                    },
                    "500": {
                        "description": "Une erreur s'est produite côté serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/movements/out/{code}": {
            "get": {
                "tags": [
                    "Sortie de stock"
                ],
                "summary": "Afficher un mouvement de sortie spécifique",
                "description": "Récupère les détails d'un mouvement OUT (sortie) avec produits, district et audit.",
                "operationId": "e1067849336fa97b5bc83f064f02eac8",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code du mouvement OUT",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "RSHXXXXX"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du mouvement de sortie récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovementDetailOut"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mouvement non trouvé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "500": {
                        "description": "Erreur interne serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sortie de stock"
                ],
                "summary": "Modifier un mouvement de sortie",
                "description": "Met à jour les informations d'un sorite de stock.",
                "operationId": "098e3daae0b6d204c967cda179ddb07b",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code du sortie de sortie",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "RSHXXXXX"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "description": "Liste des produits au format 'CODE:QTY;CODE2:QTY'",
                                        "type": "string",
                                        "example": "PR3D06:10"
                                    },
                                    "district_id": {
                                        "description": "Ville/Quartier associé au demande de sortie",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "name": {
                                        "description": "Nom du client/fournisseur",
                                        "type": "string",
                                        "example": "Nouveau nom",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "description": "Téléphone mis à jour",
                                        "type": "string",
                                        "example": "0699887766",
                                        "nullable": true
                                    },
                                    "address": {
                                        "description": "Adresse modifiée",
                                        "type": "string",
                                        "example": "Nouvelle adresse",
                                        "nullable": true
                                    },
                                    "note": {
                                        "description": "Note ou précision ajoutée",
                                        "type": "string",
                                        "example": "Mise à jour après vérification",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sortie de stock mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovementDetailOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Les données fournies sont invalides"
                    },
                    "403": {
                        "description": "Vous n'avez pas les autorisations nécessaires"
                    },
                    "500": {
                        "description": "Une erreur s'est produite côté serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sortie de stock"
                ],
                "summary": "Supprimer un mouvement de sortie",
                "description": "Supprime un mouvement de sortie de stock selon son code.",
                "operationId": "c09396304fb9816dcf2f29e42fbeb82a",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code du mouvement de sortie",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "MOVOUT00987"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mouvement de sortie supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mouvement de sortie supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Vous n'avez pas les autorisations nécessaires"
                    },
                    "404": {
                        "description": "Mouvement non trouvé"
                    },
                    "500": {
                        "description": "Une erreur s'est produite côté serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MovementListDetailOut": {
                "title": "Détail du sortie de stock",
                "properties": {
                    "district": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Casablanca - Al fida"
                            },
                            "arabic_name": {
                                "type": "string",
                                "example": "الدار البيضاء - الفداء"
                            }
                        },
                        "type": "object"
                    },
                    "code": {
                        "type": "string",
                        "example": "RSHXXXXX"
                    },
                    "status": {
                        "type": "string",
                        "example": "PENDING"
                    },
                    "fee": {
                        "type": "number",
                        "example": 0
                    },
                    "name": {
                        "type": "string",
                        "example": "test"
                    },
                    "phone": {
                        "type": "string",
                        "example": "06XXXXXXXX"
                    },
                    "address": {
                        "type": "string",
                        "example": "Test"
                    },
                    "last_action_at": {
                        "type": "string",
                        "example": "2025-01-01 00:00:00"
                    },
                    "products": {
                        "$ref": "#/components/schemas/MovementProduct"
                    }
                },
                "type": "object"
            },
            "MovementProduct": {
                "title": "Produit d'un mouvement",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "PRXXXX"
                    },
                    "reference": {
                        "type": "string",
                        "example": "Test"
                    },
                    "name": {
                        "type": "string",
                        "example": "Test"
                    },
                    "initial_quantity": {
                        "type": "integer",
                        "example": 20
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 20
                    },
                    "note": {
                        "type": "string",
                        "example": "Test",
                        "nullable": true
                    },
                    "under_quarantine": {
                        "type": "integer",
                        "example": 0
                    },
                    "active": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "MovementDetailOut": {
                "title": "Détail du mouvement de sortie",
                "properties": {
                    "district": {
                        "description": "Informations du district",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Casablanca - Al fida"
                            },
                            "arabic_name": {
                                "type": "string",
                                "example": "الدار البيضاء - الفداء"
                            }
                        },
                        "type": "object"
                    },
                    "code": {
                        "type": "string",
                        "example": "RSHXXXXX"
                    },
                    "status": {
                        "type": "string",
                        "example": "PENDING"
                    },
                    "fee": {
                        "type": "number",
                        "example": 0
                    },
                    "name": {
                        "type": "string",
                        "example": "Nom client"
                    },
                    "phone": {
                        "type": "string",
                        "example": "06XXXXXXXX"
                    },
                    "address": {
                        "type": "string",
                        "example": "Adresse client"
                    },
                    "last_action_at": {
                        "type": "string",
                        "example": "2025-12-12 11:01:30"
                    },
                    "products": {
                        "$ref": "#/components/schemas/MovementProduct"
                    },
                    "audits": {
                        "description": "Historique d’audit",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MovementAudit"
                        }
                    }
                },
                "type": "object"
            },
            "MovementAudit": {
                "title": "Audit d’un mouvement",
                "properties": {
                    "data": {
                        "description": "Données modifiées",
                        "type": "object",
                        "example": {
                            "status": "PENDING",
                            "phone": "06XXXXXXXX",
                            "address": "Casa"
                        }
                    },
                    "event": {
                        "type": "string",
                        "example": "updated"
                    },
                    "user": {
                        "type": "string",
                        "example": "Sendit vendeur"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "vendredi 12 décembre 2025, à 11:01"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}